+++ src/Components/Example.php
@@ -10,11 +10,11 @@ class Example extends BlockEditComponent
             <div class="space-y-4">
                 <div>
                     <label for="title" class="mb-1">Title</label>
-                    <input type="text" id="title" wire:model.debounce.500ms="data.title" class="w-full border border-gray-200 px-3 py-1 rounded-md">
+                    <input type="text" id="title" wire:model.live.debounce.500ms="data.title" class="w-full border border-gray-200 px-3 py-1 rounded-md">
                 </div>
                 <div>
                     <label for="content" class="mb-1">Content</label>
-                    <textarea id="content" wire:model.debounce.500ms="data.content" class="w-full border border-gray-200 px-3 py-1 rounded-md"></textarea>
+                    <textarea id="content" wire:model.live.debounce.500ms="data.content" class="w-full border border-gray-200 px-3 py-1 rounded-md"></textarea>
                 </div>
             </div>
         blade;